nix-shell -> nix develop (#12)
Convert nix/shell.nix to a root flake.nix which is more up-to-date with the nix community.

It also adds brax with the caveat that mujoco is not installed, so there will be runtime dependencies if that path is invoked. Similarly, tests were skipped, because of a failure that wasn't investigated, so expect other errors.